requirements.txt
pyTelegramBotAPI
Pillow
